+++ docs/conf.py
@@ -18,7 +18,7 @@ copyright = '2023, Sebastian Kuebeck'
 author = 'Sebastian Kuebeck'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.0'
+release = '0.1.1'
 
 
 # -- General configuration ---------------------------------------------------