+++ CHANGES.md
@@ -1,6 +1,10 @@
 # Changes
 
-## 0.1.0 (2023-09-22)
+## 0.1.1 (2023-09-22)
+- Reduced zombie process reaping CPU load
+- Fixed integration test
+- Added zombie process reaper test
 
+## 0.1.0 (2023-09-22)
 - Zombie process reaping
 - Log Collector Extension for programs that refuse to log to stdout/stderr